오류로 컴포넌트 삭제
